# perf: Improve mandelbrot/1.go