```
goos: darwin
goarch: arm64
pkg: mandelbrot
cpu: Apple M4 Pro
BenchmarkMandelbrotOrig-14           392           2968428 ns/op            7215 B/op          4 allocs/op
BenchmarkMandelbrotOrig-14           404           2937936 ns/op            7213 B/op          4 allocs/op
BenchmarkMandelbrotOpt-14            685           1749763 ns/op            7214 B/op          4 allocs/op
BenchmarkMandelbrotOpt-14            687           1753325 ns/op            7215 B/op          4 allocs/op
PASS
ok      mandelbrot      4.914s
```